fix clue checking bugs, add docstrings, move more utils to util.gd
I believe I have all the edge cases for clues now
One of the bugs was fixed by changing a pattern from X to X[_X]*
Another fix was handling 0 in clues
Final bug was a logic bug where what's now check_unambiguous_placement was ignoring parts of the line when checking if the first or last clue worked in a spot.

Hopefully the added documentation makes up for all the regex use here.